refactor(workflows): extract chat query orchestration to chat_flow
refactor(workflows): チャットクエリオペレーションをchat_flowへ分離

- Move handle_query and related orchestration logic to workflows/chat_flow.py
- Inject AppDeps to manage chain execution, background tasks, and database queries
- Keep UI event wiring and component routing inside the app.py composition root
- handle_queryと関連するオーケストレーションロジックをworkflows/chat_flow.pyに移動
- AppDepsを注入し、チェーン実行、バックグラウンドタスク、およびDBクエリを管理
- UIイベントの配線とコンポーネントルーティングはapp.py内に維持
