+++ frontend/app.py
@@ -30,6 +30,7 @@ from frontend.components.source_panel import render_sources
 from frontend.deps import AppDeps
 from frontend.state.session_state import bootstrap_session_state, reset_conversation_state
 from frontend.workflows.chat_stream import render_streamed_ai_answer
+from frontend.workflows.chat_flow import handle_query
 
 import csv
 import uuid
@@ -60,46 +61,6 @@ render_new_chat_button(reset_conversation_state)
 
 render_sidebar(deps.db_connection_factory)
     
-
-def handle_query(question):
-    with st.chat_message("user"):
-        st.markdown(question)
-
-    st.session_state.messages.append({"role": "user", "content": question})
-
-    session_id = st.session_state.conv_id
-    full_response, sources = render_streamed_ai_answer(deps.rag_chain, question, session_id)
-
-    #luu cau tra loi cua AI vao history de hien thi
-    st.session_state.messages.append({
-        "role": "ai", 
-        "content": full_response,
-        "sources": sources #avoid losing sources when reload
-    })
-
-    session_id = st.session_state.conv_id
-    user_id = st.session_state.user_id
-    message_count = len(st.session_state.messages) #khi bắt đầu có câu hỏi đầu tiên của user + answer của llm
-                                                   #tức 2 message trong state -> lập tức run task sinh title
-
-    is_first_exchange = message_count == 2
-    already_scheduled = session_id in st.session_state.title_generation_started
-
-    if is_first_exchange and not already_scheduled:
-        first_question = st.session_state.messages[0]["content"] #câu hỏi của user
-        first_response = st.session_state.messages[1]["content"] #response của llm
-        st.session_state.title_generation_started.add(session_id)
-
-
-        def _generate_and_save_title(conv_id: str, uid: str, first_q: str, first_r: str):
-            insert_title_conversations(deps.db_connection_factory(), conv_id, uid, deps.title_generator(first_q, first_r))
-
-
-        #thread sinh title độc lập với thread nhập input -> tức người dùng vẫn có thể nhập câu hỏi tiếp theo
-        #và thread sinh title chạy ngầm
-        deps.background_scheduler(_generate_and_save_title, session_id, user_id, first_question, first_response)
-
-        
 #CHI ve 1 an duy nhat - tranh loi double display
 for msg in st.session_state.messages:
     with st.chat_message(msg["role"]):
@@ -136,25 +97,22 @@ def hero_section():
         "Học phần song hành là gì"
     ]
 
-    def set_prompt(text):
-        st.session_state.prompt_trigger = text
-
     with col1:
         if st.button(suggestions[0], use_container_width=True):
-            handle_query(suggestions[0])
+            handle_query(suggestions[0], deps)
             st.rerun()
         
         if st.button(suggestions[2], use_container_width=True):
-            handle_query(suggestions[2])
+            handle_query(suggestions[2], deps)
             st.rerun()
 
     with col2:
         if st.button(suggestions[1], use_container_width=True):
-            handle_query(suggestions[1])
+            handle_query(suggestions[1], deps)
             st.rerun()
         
         if st.button(suggestions[3], use_container_width=True):
-            handle_query(suggestions[3])
+            handle_query(suggestions[3], deps)
             st.rerun()
 
 if not st.session_state.messages:
@@ -167,14 +125,14 @@ if not st.session_state.messages:
 if "prompt_trigger" in st.session_state:
     prompt = st.session_state.prompt_trigger
     del st.session_state.prompt_trigger
-    handle_query(prompt)
+    handle_query(prompt, deps)
     st.rerun()
 
 
 #Roi moi kiem tra chat input UI (uu tien 2)
 #placeholder bang tieng viet
 elif prompt := st.chat_input("Nhập câu hỏi về quy chế, hoặc chat chit..."):
-    handle_query(prompt)
+    handle_query(prompt, deps)
     st.rerun()
 
 render_feedback()
